AdvanceTradingBot/requirements.txt
websockets==11.0.3
numpy==1.24.3
pandas==2.0.3
python-binance==1.0.19
pyyaml==6.0
asyncio==3.4.3
scikit-learn==1.3.0
statsmodels==0.14.0
